+++ src/main/java/net/geforcemods/securitycraft/items/ModuleItem.java
@@ -1,5 +1,6 @@
 package net.geforcemods.securitycraft.items;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import net.geforcemods.securitycraft.ClientHandler;
@@ -10,10 +11,10 @@ import net.geforcemods.securitycraft.api.LinkableBlockEntity;
 import net.geforcemods.securitycraft.inventory.DisguiseModuleMenu;
 import net.geforcemods.securitycraft.inventory.ModuleItemContainer;
 import net.geforcemods.securitycraft.misc.ModuleType;
-import net.geforcemods.securitycraft.util.ModuleUtils;
 import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
+import net.minecraft.nbt.StringTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
@@ -33,6 +34,7 @@ import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.scores.PlayerTeam;
 import net.minecraftforge.fmllegacy.network.NetworkHooks;
 
 public class ModuleItem extends Item {
@@ -137,7 +139,7 @@ public class ModuleItem extends Item {
 				affectsEveryone = tag.getBoolean("affectEveryone");
 
 				if (!affectsEveryone) {
-					playerCount = ModuleUtils.getPlayersFromModule(stack).size();
+					playerCount = ModuleItem.getPlayersFromModule(stack).size();
 					teamCount = tag.getList("ListedTeams", Tag.TAG_STRING).size();
 				}
 			}
@@ -171,5 +173,30 @@ public class ModuleItem extends Item {
 
 	public boolean canBeCustomized() {
 		return canBeCustomized;
+	}
+
+	public static boolean doesModuleHaveTeamOf(ItemStack module, String name, Level level) {
+		PlayerTeam team = level.getScoreboard().getPlayersTeam(name);
+
+		//@formatter:off
+		return team != null && module.getOrCreateTag().getList("ListedTeams", Tag.TAG_STRING)
+				.stream()
+				.filter(tag -> tag instanceof StringTag)
+				.map(tag -> ((StringTag) tag).getAsString())
+				.anyMatch(team.getName()::equals);
+		//@formatter:on
+	}
+
+	public static List<String> getPlayersFromModule(ItemStack stack) {
+		List<String> list = new ArrayList<>();
+
+		if (stack.getItem() instanceof ModuleItem && stack.hasTag()) {
+			for (int i = 1; i <= MAX_PLAYERS; i++) {
+				if (stack.getTag().getString("Player" + i) != null && !stack.getTag().getString("Player" + i).isEmpty())
+					list.add(stack.getTag().getString("Player" + i).toLowerCase());
+			}
+		}
+
+		return list;
 	}
 }